415 Introduces useCheckEngineStatus hook to check the status and showing the error message in case the engine is not running
