+++ packages/orchestrator-ui-components/src/components/WfoPageTemplate/WfoSidebar/WfoStartCreateWorkflowButtonComboBox.tsx
@@ -3,13 +3,8 @@ import React from 'react';
 import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/router';
 
-import { ToastTypes } from '@/contexts';
 import { GET_WORKFLOWS_FOR_DROPDOWN_LIST_GRAPHQL_QUERY } from '@/graphqlQueries/workflows/workflowsQueryForDropdownList';
-import {
-    useEngineStatusQuery,
-    useQueryWithGraphql,
-    useToastMessage,
-} from '@/hooks';
+import { useCheckEngineStatus, useQueryWithGraphql } from '@/hooks';
 
 import {
     WfoButtonComboBox,
@@ -20,9 +15,7 @@ import { PATH_START_NEW_PROCESS } from '../paths';
 export const WfoStartCreateWorkflowButtonComboBox = () => {
     const router = useRouter();
     const t = useTranslations('common');
-    const tErrors = useTranslations('errors');
-    const { isEngineRunningNow } = useEngineStatusQuery();
-    const toastMessage = useToastMessage();
+    const { isEngineRunningNow } = useCheckEngineStatus();
 
     const { data } = useQueryWithGraphql(
         GET_WORKFLOWS_FOR_DROPDOWN_LIST_GRAPHQL_QUERY,
@@ -51,21 +44,13 @@ export const WfoStartCreateWorkflowButtonComboBox = () => {
     const handleOptionChange = async (
         selectedProduct: WorkflowComboBoxOption,
     ) => {
-        const isEngineRunning = await isEngineRunningNow();
-
-        if (!isEngineRunning) {
-            return toastMessage.addToast(
-                ToastTypes.ERROR,
-                tErrors('notAllowedWhenEngineIsNotRunningMessage'),
-                tErrors('notAllowedWhenEngineIsNotRunningTitle'),
-            );
+        if (await isEngineRunningNow()) {
+            const { workflowName, productId } = selectedProduct.data;
+            router.push({
+                pathname: `${PATH_START_NEW_PROCESS}/${workflowName}`,
+                query: { productId },
+            });
         }
-
-        const { workflowName, productId } = selectedProduct.data;
-        router.push({
-            pathname: `${PATH_START_NEW_PROCESS}/${workflowName}`,
-            query: { productId },
-        });
     };
 
     return (
