+++ packages/orchestrator-ui-components/src/components/WfoStartTaskButtonComboBox/WfoStartTaskButtonComboBox.tsx
@@ -4,13 +4,8 @@ import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/router';
 
 import { PATH_START_NEW_TASK } from '@/components';
-import { ToastTypes } from '@/contexts';
 import { GET_WORKFLOWS_FOR_DROPDOWN_LIST_GRAPHQL_QUERY } from '@/graphqlQueries/workflows/workflowsQueryForDropdownList';
-import {
-    useEngineStatusQuery,
-    useQueryWithGraphql,
-    useToastMessage,
-} from '@/hooks';
+import { useCheckEngineStatus, useQueryWithGraphql } from '@/hooks';
 
 import {
     WfoButtonComboBox,
@@ -20,9 +15,7 @@ import {
 export const WfoStartTaskButtonComboBox = () => {
     const router = useRouter();
     const t = useTranslations('common');
-    const tErrors = useTranslations('errors');
-    const { isEngineRunningNow } = useEngineStatusQuery();
-    const toastMessage = useToastMessage();
+    const { isEngineRunningNow } = useCheckEngineStatus();
 
     const { data } = useQueryWithGraphql(
         GET_WORKFLOWS_FOR_DROPDOWN_LIST_GRAPHQL_QUERY,
@@ -45,20 +38,12 @@ export const WfoStartTaskButtonComboBox = () => {
     const handleOptionChange = async (
         selectedProduct: WorkflowComboBoxOption,
     ) => {
-        const isEngineRunning = await isEngineRunningNow();
-
-        if (!isEngineRunning) {
-            return toastMessage.addToast(
-                ToastTypes.ERROR,
-                tErrors('notAllowedWhenEngineIsNotRunningMessage'),
-                tErrors('notAllowedWhenEngineIsNotRunningTitle'),
-            );
+        if (await isEngineRunningNow()) {
+            const { workflowName } = selectedProduct.data;
+            router.push({
+                pathname: `${PATH_START_NEW_TASK}/${workflowName}`,
+            });
         }
-
-        const { workflowName } = selectedProduct.data;
-        router.push({
-            pathname: `${PATH_START_NEW_TASK}/${workflowName}`,
-        });
     };
 
     return (
