+++ packages/orchestrator-ui-components/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useCheckEngineStatus';
 export * from './useQueryWithGraphql';
 export * from './useEngineStatusQuery';
 export * from './useMutateProcess';
